backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
openai==1.57.4
zhipuai
supabase==2.10.0
twilio==9.3.7
boto3==1.35.85
botocore==1.35.85
gspread==6.1.4
google-auth==2.37.0
httpx<0.28,>=0.26
python-multipart==0.0.20
aiofiles==24.1.0
asyncio-throttle==1.0.2
structlog==24.4.0
